Redesigned EquationBase#10
This is a major change that EquationBase will realize what
was suggested: being a general equation handler by giving a name.

One of the major changes that will suit in the new structure
is PETSc matrices and vectors for different equation will live
in different instances of derived classes of EquationBase.