Updates for compatibility with deal.II version 9.0.0:
deal.II defines a global Assert() macro and gmock has an internal function called Assert(), which is bad.
Added test_helpers/gmock_wrapper.h to include gmock.h without the preprocessor mistaking that function definition for a macro.
This issue only happened to manifest with the new version of deal.II because a dependency on deal.II/base/exceptions.h was added to deal.II/base/mpi.h.
If we could guarantee that gmock-internal-utils.h is included before exceptions.h, this wouldn't be an issue,
nor if we could guarantee that we could #undef Assert before including gmock.h without breaking anything in deal.II,
but this ugly wrapper header with copy/pasted code that would need to be updated if deal.II's implementation of the Assert() macro changes is the best solution I could come up with without modifying deal.II itself.